Revert "LPS-166106 Avoid the use of static classes. To do that, move inner classes to a separate class, so they are accessible from the method .newInstance() in the GraphQL infrastructure."
This reverts commit 502820cb5df9377719476250bf7cd78ba1d80532.
