+++ modules/apps/portal-vulcan/portal-vulcan-test/src/testIntegration/java/com/liferay/portal/vulcan/internal/graphql/servlet/test/GraphQLServletTest.java
@@ -56,7 +56,7 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 				ServletData.class, testServletData, null);
 
 		String testDTOSimpleClassName = StringUtil.lowerCaseFirstLetter(
-			TestQuery.TestDTO.class.getSimpleName());
+			TestDTO.class.getSimpleName());
 
 		GraphQLField graphQLField = new GraphQLField(
 			testDTOSimpleClassName, new GraphQLField("field"),
